requirements.txt
blinker   
click      
colorama  
Flask     
itsdangerous 
Jinja2    
MarkupSafe
pip       
setuptools  
Werkzeug   
wheel    
certifi
chardet
charset-normalizer
colorama
et-xmlfile
filelock
fsspec
googletrans
h11
h2
hpack
hstspreload
httpcore
httpx
hyperframe
idna
Jinja2
keyboard
llvmlite
MarkupSafe
more-itertools
mpmath
networkx
numba
numpy
openai-whisper
openpyxl
pandas
PyAudio
pydub
python-dateutil
pytz
regex
requests
rfc3986
six
sniffio
sympy
tiktoken
torch
tqdm
typing_extensions
tzdata
urllib3
gunicorn